Allow error logger to write events to file (#11944)
Part of
[AB#1847](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1847)

As a part of
https://github.com/microsoft/FluidFramework/pull/11928/files#

We want to log all the errors to the file, the `InactiveObjectX` errors
are also logged to a separate file for easy browsing. Another considered
solution would be to run the tests in Mocha with Fluid in verbose mode
and then to store those in a file.

This would be the effort to store events in a file. The `nyc` directory
was used as `nyc` was ignored by git.