Fix the Missing Deltas when Switch the Cluster (#11919)
When we switch clusters, it would have the GetDeltas_Exception. We can
reproduce the following scenario:

1. Cluster A opens a new session, and adds ops [1, 2, 3, 4] in Cluster
A's cosmos DB. We have the client summary at seq 2. The session ends and
makes the service summary = the client summary at seq 2 + [3, 4].
2. Cluster B opens the session, and adds ops [5, 6] in Cluster B's
cosmos DB. It would read the correct service summary from step 1. We
don't have a client summary. However, when the session ends, the new
service summary = client summary at seq 2 + [5, 6]
3. Cluster A opens a new session again, it would read the wrong service
summary from step 2. Meanwhile, it would request ops after the last
client summary at seq 2, and there are missing operations [5, 6] that
are in cluster B's cosmos DB.


To fix this bug, when we fetch the logtails, we would fill in the
missing operations from the last summary messages.

We make the session stickiness time as cluster-related variable in this
pr.