Fix index summary loading in SharedTree (#11960)
SharedTree's indexes currently allow for custom loading, but they are
not able to actually retrieve blobs from the storage service without
knowing the path element above them. This PR "scopes" the service object
passed to each index with the path preemptively so that each index can
work within a local namespace and not worry about collisions with other
indexes. Also:

* Add tests for the loading of indexes and summarizing of SharedTree
* Stub out SchemaIndex loading so that it doesn't cause failures when
testing summarization
* Upgrade `MockObjectStorageService` to respect summary elements that
are more than one-deep in a summary tree.
* Factor out some test utilities in sharedTreeCore.spec.ts