Update repo-policy-check with upcoming namespace guidelines (#11853)
## Description

This updates the repo policy check tooling to account for some upcoming
changes to our internal publishing guidelines. Generally, the new
guidelines distinguish between packages that should only ever be used by
the fluidframework team and those that may be used by other teams
internally at Microsoft.

Contributors to this repo are generally free to make breaking changes to
packages in the former category, since the set of infrastructure/other
packages that may depend on them is known.

Since this loosens the policy check constraints on which packages may be
published internally, I also strengthened the check that we don't
accidentally depend on private packages from non-private packages to
actually perform the computation. I ended up just doing this in a second
pass of reading the files, though with some more involved refactoring we
could instead process the packages in topological order to solve that
problem.

Co-authored-by: Abram Sanderson <absander@microsoft.com>