GC will not run if there if no messages have ever been processed for a container (#11957)
## Description
This change does the following:
- It moved the initialization of GC base state for non-summarizer
clients to when the client becomes active (its in write mode and
connected). This will ensure that the client will see its own join op
which means all trailing ops would have been processed and the timestamp
>= of the join op will be used for initialization of GC state.
-It removes the code that handles scenarios where there is no current
reference timestamp for GC to work with.

GC needs the current reference timestamp for 2 reasons:
- To add the time when an object was found to be unreferenced.
- To find how long an object has been unreferenced for. This will help
in determining when to delete an object.

For the current timestamp, we need a server generated timestamp as the
client's clock can be skewed / incorrect and its hard to keep it sync'd
across all clients. We use the timestamp of the last message that was
processed as the current reference timestamp. The last message is also
written into the summary and is read when a container loads with this
summary.
If there is no current timestamp, GC will not run. This can happen in
the following 2 scenarios:
- GC is called when there have not been any ops processed for a
container and it is in read mode. In write mode, the container processes
its own join op so its timestamp can be used. There is no point in
running GC here anyway because references in the container change only
via ops. Also, summarizer client that runs GC always does so in write
mode.
- GC is called when a container loads from a summary generated by an
older client (runtime version <0.47.0) and is in read mode. In versions
<0.47.0, the last message was not written in the summary so when the
container loads, it does not find the last message in the summary.
Again, this scenario is not interesting because summarizer client that
runs GC always does so in write mode.