Close the container if SweepReady node is used (Behind feature gate, for dogfood use only) (#11963)
To help us and our partners flush out bugs with how datastores are
referenced, this change adds an option which will close the container
when an object is used after being marked as "SweepReady". This means
the Sweep Timeout has elapsed since it was marked as unreferenced.
Closing the container at this time will cause some user pain - users
will be locked out of their documents (see follow-up PR for a way to
ease that pain) - but it's only to be enabled in Dogfood, where users
can reach out and help our teams get traction on these bugs. Until all
such bugs are fixed, turning on Sweep represents a Data Corruption
liability.

Part of AB#1857