[build-tools] Add tests for matching success/failure actions in release state machine (#12010)
In most cases, a missing failure or success action is a bug. It's a
mistake that's easy to make, so this change adds tests to check that
states with success/failure actions have both. The exception is the Do*
actions.

Fixing this test fixed a bunch of bugs in the state machine due to
copy/paste errors, refactoring, etc.