Add support for only closing container once per X time period due to SweepReady Object Used (#12004)
To help us and our partners flush out bugs with how datastores are
referenced in Dogfood environments, #11963 added an option which will
close the container when an object is used after being marked as
"SweepReady". This means the Sweep Timeout has elapsed since it was
marked as unreferenced.

Closing the container at this time will cause some user pain - users
will be locked out of their documents. So this change adds a mechanism
using localStorage by which we can configure a time period during which a
given document will only close once with this error.

This will allow us to balance the user pain v. usefulness from
diagnostic standpoint to enable our partners and us to track down the
remaining sweep-blocking bugs - Until all such bugs are fixed, turning
on Sweep represents a Data Corruption liability.