Odsp binary snapshot parsing: Rework string manipulation code (#12047)
More clearly separate blobs, dictionary strings and strings in binary snapshot format parsing code.
Reduce amount of time we parse blobs when dealing with dictionary strings (by keeping a dictionary of strings, not blobs), and overall attempt to reduce number of objects we create on a string parsing path.
Removed a bunch of validation in getAndValidateNodeProps() and push it to callers, as they already validate presence and type of properties (so amount of code run is reduced). Returning extra properties is Ok here.

Removing ISnapshotTreeEx - there is no good reason to create empty commits everywhere.