[v2int/1.1] Restore bindToContext functionality (#12048)
After discussing the challenges being faced by our internal partners
integrating the 2.0.0-internal.1.1.x release due to the removal of
`bindToContext` (see #10627), we're adding it back in. They are tracking
the work required to fully migrate off the request pattern to using
handles, which will render this fully obsolete and at that point we will
remove it.

## Does this introduce a breaking change?

No... it undoes a breaking change, such that this change is additive.
Typically I think we avoid even backwards-compatible API changes in
patch releases, but all interfaces being modified are for internal use
only (we don't expect partners to be implementing them). And this is
required to unblock our partner integrating the release.