Fix Race Condition in Initalize Protocol Tree (#12083)
## Description

There is a race condition in the initialization of protocol tree that
can lead to initial clients being set after protocol tree is created,
but not set due to the async nature of the how we initialize the
protocol tree.

Specifically, we kick off initializeProtocolState as an async task. The
task runs, so the protocol tree object is created, but
this._protocolHandler is not set yet. Before the task returns we
connect, so the delta manager connect handler runs, it sees
this._protocolHandler is undefined, and puts initial clients in
this._initialClients. Then initializeProtocolState completes, and
this._protocolHandler is set, without ever looking at
this._initialClients which leads to lost member of the audience.

We are seeing this consistently fail in our test runs for intv2/1.1:
https://github.com/microsoft/FluidFramework/pull/12071