[GC-Stress] Enable shrinking SweepTimeout for testing (#12092)
This is WIP version of #12044. I'm pausing that work in terms of merging
to main until next week.

This way we can shrink Session Expiry and the Sweep Timeout in the
stress test and watch for SweepReady errors.

Here are the settings to set:

* `Fluid.GarbageCollection.TestOverride.SessionExpiryMs`
* `Fluid.Driver.Odsp.TestOverride.DisableSnapshotCache` (disables
snapshot caching in driver and uses 0 for the Sweep Timeout calculation)
* `Fluid.GarbageCollection.TestOverride.SweepBufferMs`

Then the total SweepTimeout will be `sessionExpiryMs + sweepBufferMs`