Increased the delay between ops for each client (#12110)
Adjusted the op rate for each client by dividing total op rate / number
of clients. This will increase the delay between ops in each client and
the overall op rate would still be total op rate.

Note that each client has multiple data stores which are each sending
ops. So, the opRatePerDataStore would be opRatePerClient /
dataStoreCount. Will do this as a follow up when we put a cap on the
number of data stores.