Added timestamp to GarbageCollection event after GC runs. (#12138)
Added the current reference timestamp to the GarbageCollection event.
This is the timestamp used by GC to mark the time when a data store
becomes unreferenced. It will help in debugging inactive object errors
where we can correlate in which GC run a data store was marked as
unreferenced based on its age (how long it has been unreferenced for).