[main > release/0.59.4000]: Fix 404 error during HandleSummaryAckError error (#11336)  (#12147)
## Description

This handle the flow where while summarizer is running into handleAcks
flow and waiting for ack promise to get resolved, the summarizer can
make a new summary. Now when we try to refresh the state from handles of
previous ack, and try to fetch snapshot corresponding to that it might
not be present in cache and server(server deletes previous snapshots)
and we get 404 from server and run into "HandleSummaryAckError". So
before giving up, we try one more time with refreshing the state from
latest snapshot handle.

Co-authored-by: NicholasCouri <nichoc@microsoft.com>