Handle cluster expansion case in IdCompressor (#12017)
This fixes a bug in which a client of the IdCompressor would receive
incorrect (specifically, duplicate) IDs after decompression. The
compressor did not properly update the session normalizer to account for
outstanding local IDs (i.e. local IDs that had not yet been finalized)
when performing the cluster expansion operation.