Add --all argument to release report command (#12122)
**Adds a `--all` flag to list all the releases of a particular release
group or package.** This is useful when there are a lot of releases over
a short period of time, or if you're trying to figure out what the last
release of a particular package was.

**Adds the version scheme and dependency ranges to the full JSON
output.** This makes it easier for partners to integrate releases
without using the version tool separately.

**`--output` argument is now a directory.** When specified, multiple
report files will be output to it. There are four report formats:
simple, full, caret, and tilde.

- Simple is a map of package to versions. This is what we have
traditionally provided to partners.
- Full includes the all metadata about the package.
- Caret and tilde are maps of package names to the semver range that
should be used for either `^` or `~` dependencies. These can be consumed
by partners that want to use those dependency ranges directly. Note that
for packages using semver, the `^` or `~` will be prepended, while Fluid
internal versions will use `>= <` ranges. The strings are meant to be
used as-is.

Co-authored-by: Joshua Smithrud <54606601+Josmithr@users.noreply.github.com>