[build-tools] Generate lockfile correctly and simplify code (#12157)
When trying to switch the CI system to use `flub generate packageJson`,
I noticed that the lockfiles were wrong. After investigating, I found
some inverted logic in the code that was copied from the old
build-tools. Rather than debug further, I opted to remove the copied
code, move the existing code into a library file, and then export it so
it can be called directly from the command.