[build-tools] Fix links to command source in README (#12155)
The links to the source for commands in the build-tools CLIs is wrong. I
originally thought that fixing it would require our own readme command,
overriding the one that oclif provides. Turns out that's not strictly
necessary if I change the output folders for the packages to `lib` from
`dist`. Looks like oclif special-cases the lib path:

https://github.com/oclif/oclif/blob/18b75cfbe1bf6a0e076709410aedf75e3e8f843a/src/commands/readme.ts#L225

I just moved stuff to lib since it seemed easier. But I also left the
command I implemented, `generate readme`. The version I added doesn't
output the platform/architecture info into the readme, because that
causes unnecessary churn on the readme when individual devs build on
their own machines.