Various fixes to dice roller tutorial (#12026)
Fixes https://dev.azure.com/fluidframework/internal/_workitems/edit/1220

The description of that bug is pasted here:

> Consider using "Note" formatting rather than "*" for AFR note
Code snippets in tutorial are not exact match to app.js in repo, e.g.:
const root = document.getElementById("root")
const root = document.getElementById("content");
Consider explaining that the "attach" call is what actually publishes to
the server (not just getting an ID)
renderDiceRoller call is not explained in the create/load sections - is
there a way to either defer showing this, or else explain it a bit in
these sections?
Clarify that using the URL hash to control create/load is just a trick
for the demo, and that a real application may have other logic here.
I'm not sure that showing the non-Fluid version of the view first is
super helpful? It's certainly interesting once you know the Fluid
version too so you can compare/contrast, but might be confusing why
we're showing a non-Fluid implementation.