Keep references on unacked removed segments (#12190)
## Description
This change reduces complexity around how we handle references on
removed segements. Specifically on unacked removes we will just leave
the references in place, and then only on acknowledgment will we manage
them based on their reference type. I've also encapsulated some of the
logic around how reference types are handled, so it is less distributed
in the code.


[AB#1845](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1845)