[build-tools] Fix lint errors (#12191)
This PR is mostly updating the build-tools lint config to align more
with our shared config.

One thing I feel a little guilty about is the overuse of
non-null-assertions, especially in the state handling code. That code
has a "data" object that gets passed around that in 99.9% of cases is
going to have the proprties defined that the state handler needs. I
found checking for undefined explicitly made the code less readable, for
no practical benefit.

I'm open to alternatives to this pattern for a follow up PR!