Use GetFileById() instead of GetFileByUrl() to get sharing information for ODSP file (#12274)
## Description

This change is needed to ensure that odsp-driver can get file sharing
information when used in Consumer scenarios. Currently we rely on
GetFileUrl() method which does not work for Consumer because it does not
support webDavUrl in Consumer format. The solution is to use
GetFileById() method which works uniformaly for Consumer and Commercial
segments

## Does this introduce a breaking change?

No breaking change, only internal implementation changes.

Co-authored-by: REDMOND\andreize <andreize@microsoft.com>