feat(build-tools): Support pnpm-workspace.yaml as workspace source (#12252)
fluid-build loads workspaces from lerna.json or package.json. This
change adds support for pnpm's `pnpm-workspace.yaml`. Versions will
still be loaded from lerna.json, but the workspace definition will be
loaded from pnpm-workspace.yaml if it's present.

This is a quality of life fix for folks who are experimenting with pnpm
in parts of the repo.