Tree: Add symbolIsFieldKey (#12305)
## Description

Adds `symbolIsFieldKey` which is necessary for APIs like EditableTree
which need to interpret attempts to access or modify properties keys
with symbols as either field access or not. Since the set of symbols
code might test an object for is unknown, it's impossible to implement
such APIs with a way to tell if a symbol is a field key.