Update the Garbage Collection (GC) documents as per the change from root data store -> aliasing (#12257)
There were recent changes to root data stores where the only way to make
a data store as root is by aliasing it. Updated the
`garbageCollection.md` document to reflect this change. Also, tried to
organize it a little better to reflect the reasons behind doing GC and
what are the responsibilities of consumers of Fluid framework.