fix blob GC test (#11101)
This was caused by #10621, which added support for blob handles using
local IDs while offline. Due to a race condition, the container would
not always be connected and uploadBlob() would return a local ID handle
instead of one using the storage ID as the test expected, which would
cause it to fail.