[r11s] Making summary reader use 'latest' summary id (#12384)
## Description

`SummaryReader` is responsible for reading the last summary status
available for a given document. That latest status is usually available
in `Historian`'s cache. However, we noticed that SummaryReader would
usually bypass the cache when fetching a summary in the context of whole
summaries. After some investigation, the reason is because
`SummaryReader` provides an explicit summary ID/SHA when sending the
request. For whole summaries, data is only cached for the latest
summary, which uses the ID `latest` as an alias.

We probably did not use `latest` in `SummaryReader` for whole summaries
because we were still working on the whole summary story when we worked
on `SummaryReader` in
https://github.com/microsoft/FluidFramework/pull/7283. Unlike its
counterpart `SummaryWriter`, which needs to obtain the explicit ref for
a given document so it can compare it to the content of the client
summary proposal, `SummaryReader` only cares about the latest piece of
data.

Therefore, this PR makes it so that `SummaryReader` requests to
`Historian` use the magical keyword `latest`. Besides not having to go
to storage due to potentially resolving the request from `Historian`'s
cache, an added benefit is that this strategy also removes the extra
call to Historian and then storage to fetch the ref.

Finally, as a fallback, if `latest` does not work, `SummaryReader` will
default to using the standard `GET ref`/`GET summary` setup.

## Does this introduce a breaking change?

No.

## Testing

Validated the setup with Clicker running the following experiment:
1. Open a clicker window and let a client summary happen
2. Close the clicker window and let a service summary happen
3. Restart Scribe in Docker
4. Reopen the clicker window, using the same URL from 1
5. Checked Historian logs confirming HTTP request sent from
`SummaryReader` was resolved from the cache, and confirmed on r11s logs
that reading the summary succeeded

## Next steps
I also added a new export to `services-client`: `LatestSummaryId`, so
that we can uniformly reference that constant in r11s, Historian and
GitRest. When a new pre-release or release of r11s server packages are
available, we can start using `LatestSummaryId` in other places as well.