Cursor (legacy): get field length (#12385)
This PR is needed for EditableTree sequence implementation as it allows
to get a length of a sequence field (which is just a number of sibling
nodes) when already navigated within i.e. `cursor.length(key)` won't
work as we already within the key.

Note that this changes only "legacy" cursors. New cursors already let
the user to get a field length with `getFieldLength`.