Turn SummaryStatus:Behind into a regular event and summarizerClientElection to be false in the stress tests. (#12452)
The SummaryStatus:Behind event is dependent on the number of clients in
the quorum, which is currently up to 120 clients during the stress test
runs. Taking in consideration we create incidents based on number of
errors (currently > 50), a single second in which the clients get behind
will trigger up to 120 errors during the stress test runs. Even in
production it might skew out our telemetry data. After talking to the
OCEs ( Navin, Si) and Jatin, we ve decided to move it to a regular
telemetry event for now.

Also, we are taking the chance to enforce summarizerClientElection to be
always false during the stress test runs.

Sample incidents: 

https://portal.microsofticm.com/imp/v3/incidents/details/340813540/home

https://portal.microsofticm.com/imp/v3/incidents/details/341118841/home