Turn SummaryStatus:Behind into a regular event and summarizerClientEl… (#12456)
…ection to be false in the stress tests. (#12452)

The SummaryStatus:Behind event is dependent on the number of clients in
the quorum, which is currently up to 120 clients during the stress test
runs. Taking in consideration we create incidents based on number of
errors (currently > 50), a single second in which the clients get behind
will trigger up to 120 errors during the stress test runs. Even in
production it might skew out our telemetry data. After talking to the
OCEs ( Navin, Si) and Jatin, we ve decided to move it to a regular
telemetry event for now.

Also, we are taking the chance to enforce summarizerClientElection to be
always false during the stress test runs.

Sample incidents: 

https://portal.microsofticm.com/imp/v3/incidents/details/340813540/home

https://portal.microsofticm.com/imp/v3/incidents/details/341118841/home

For more information about how to contribute to this repo, visit [this
page](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

For specific guidelines for Pull Requests in this repo, visit [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

The sections included below are suggestions for what you may want to
include.
Feel free to remove or alter parts of this template that do not offer
value for your specific change.

## Description

> A concise description of the changes (bug or feature) and their
impact/motivation.
> If this description is short enough to be used as the title, delete
this section and just use the title.

> For bug fixes, also include specifics of how to reproduce it / confirm
it is fixed.

> If this Pull Request should close/resolve any issues when merged, use
[the special syntax for
that](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword)
here.

## Reviewer Guidance

> List any specific things you want to get reviewer opinions on, and
anything a reviewer would need to know to review this PR effectively.
> Things you might want to include:
>
> - Questions about how to properly make automated tests for your
changes.
> -   Questions about design choices you made.
> - Descriptions of how to manually test the changes (and how much of
that you have done).
> -   etc.
>
> If you have any questions in this section, consider making the PR a
draft until all questions have been resolved.

## Does this introduce a breaking change?

> If this introduces a breaking change, please describe the impact and
migration path for existing applications below.
> See
[Breaking-vs-Non-breaking-Changes](https://github.com/microsoft/FluidFramework/wiki/Breaking-vs-Non-breaking-Changes)
for details.
> If there are no breaking changes, delete this section.

## Any relevant logs or outputs

> - Use this section to provide either screenshots or output of logs as
code snippets

## Other information or known dependencies

> - Any other information or known dependencies that is important to
this PR.
> -   Links to internal bug/work tracker items.
> -   TODO that are to be done after this PR.