Add the Fix for Missing Ops Log, and telemetry for the Ops from the Last Service Summary (#12295)
We use this pr to fix the wrong error logs that keep popping out for 
Missing ops in the fullLogTail: ...

When we get the fullLogTail, it will still pop out the error message. 
For example, if from = 5, and to = 9, the fullLogTailSequenceNumber =
[5,6,7,8,9,10], it would return [6,7,8] as missing ops, which is not
correct. This pr is to fix this case.

Meanwhile, we would print the SN for the missingOpsSN for ops fetched
from the last summary.