Cancel join session refresh timer once container is closed (#12476)
## Description

Fix for:
https://dev.azure.com/fluidframework/internal/_workitems/edit/2179?src=WorkItemMention&src-action=artifact_link
So we refresh the join session after every few mins. Now when the
container is closed, we have a refresh scheduled. I have code to clear
this referesh on "disconnect" event which seems to be not triggering.
Seems like I have to clear the refresh in dispose() method so that we
don't refresh the join session once the container is closed.