feat(build-cli): Add `merge info` command (#12437)
Adds a new command to build-cli, `merge info`. This command provides
info about the merge status of branches in the repo. The most obvious
use-case is to check how far next is behind main, so those are the
default arguments.

The output is just the number of commits behind one branch is from the
other. More details can be included by passing the `--json` flag, which outputs
a JSON object like this:

```json
{
  "branches": [
    "main",
    "next"
  ],
  "commits": [
    "859791294230b85e802510735c5798689c989d80",
    "c06be353e6046ac8e50ab04d0021d11174e500ae"
  ],
  "commitCount": 2
}
```