build: Pass build-tools version correctly (#12485)
The previous change in #11670 was incomplete. It did not properly pass
through the build tools location from the ADO UI all the way to the
right places. This change corrects that.

I tested by manually running a build with the "next" value for the build
tools source, and verified that the prerelease (tagged `@next`) version
was installed.