Azure scenario tests/runner (#12222)
Through @fluidframework/azure-scenario-runner we are exploring
improvements allowing us to build, execute and validate azure-client
scenarios in more flexible and effective manner. At this point in time,
we will mostly use the scenario runner with FRS team to validate service
characteristics. However, as we learn further, we should be able to use
it beyond service validation.