refactor(build-tools): Include ".generated" in typetest filenames (#12484)
This change makes generated files clear from their file name so reviewers
can see the generated file warning when reviewing deltas in PRs. This
change adopts the `.generated.*` convention from
<https://github.com/microsoft/TypeScript/wiki/Coding-guidelines#components>.

The fix I chose was dead-simple -- but in looking at the code I'd really
like to clean the logic up more. Still, this change is simple and
effective as-is.