webpack config cleanup (#10865)
## Description

A few webpack config changes for examples:

- remove unneeded require.resolve from webpack loaders. Webpack can
resolve these by itself just fine, and in many other places we already
rely on it doing so.
- remove unnneded quoting and escaping for env in a few package,json
files
- fix detection of production vs development in config files:
  - Simplify logic using `?.`
- Correct several configs which always returned development because
detection of production checked environment object against string. The
old incorrect behavior was determined by debugging the smde example and
checking that it worked incorrectly before and is fixed afterward.