Fix back compat for idleTime (#12539)
For more information about how to contribute to this repo, visit [this
page](https://github.com/microsoft/FluidFramework/blob/main/CONTRIBUTING.md).

## Description

[PR#10179](https://github.com/microsoft/FluidFramework/pull/10179)
introduced back-compat issue with the `idleTime` property in summary
heuristics. The property was deprecated, but no functionality was added
to continue temporary support for the property before it was officially
removed.