fix(build-cli): Require versionConstraint flag in generate:typetests (#12582)
The generate:typetests command was incorrectly calculating the previous
version when running in generate only mode. There was also not a way to
indicate what version style to use, since the flag was exclusive with
the `--generate` flag.

This change makes the `-s` flag required, so the user has to choose a
style explicitly.