Enabling op reentrancy enforcement at the DDS/runtime level (#12326)
See **ADO:2190** and **ADO:1834** for details.

We should not be allowing ops to be submitted while processing another op. This can result in inconsistencies, as the ops may end up having a different reference sequence number which is symptom that the op was produced without having a consistent view of the current state. This may lead to either different clients eventually reaching inconsistent states or to ops being applied FIFO on a single client, changing the expected order of events.

This change will allow (opt-in):
- block submitting ops in the event handler via a new runtime API which throws if a new op is being submitted while the event handling is still executing
- asserts that all ops within a JS turn batch will have the same reference sequence number. If that's not the case, the container will close and the batch will be discarded, in order to minimize the spread to the other clients
- The `AgentScheduler`, which was using this pattern, will now push the task reassignment (which was done within the event handler) to the next batch by forcing a JS turn

To enable, use the `IContainerRuntimeOptions.enableOpReentryCheck` property. The new behavior of SharedObject and the runtime can be disabled at runtime via a feature gate named `Fluid.ContainerRuntime.DisableOpReentryCheck`