fix(build-cli): Exclude independent packages when bumping release groups
When bumping deps scoped to a single release group, the `bump deps` tool was including independent
packages. This change corrects that behavior so that only the release group specified will be
updated.

BREAKING CHANGE: The `-p` flag has been changed to specify a package name, which is consistent with
other commands. Use `--prerelease` to replace former uses of `-p`.