feat(build-cli): Add handler exclusion and listing to check policy command (#12648)
Adds an `--excludeHandler` flag to the `check policy` command that can
be used to exclude one or more handlers by name. The main use-case for
this is to run policy check with or without the assert-shortcodes
handler more easily. Technically it's possible to exclude a single
handler using the regex argument, but this is much simpler.

Also, there is a new `--listHandlers` flag that will list all the
handlers including their names, so one can now query the system for
handlers, then exclude those handlers by name easily.