Implement setting tombstone (#12259)
Implements detecting and setting datastores as tombstone. Tombstoned datastores cannot send and receive ops. 
[AB#2128](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2128)

[AB#403](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/403)

Allows datastores to be marked as tombstone. 

Note:
- This PR https://github.com/microsoft/FluidFramework/pull/12293 was an
attempt to do the work without touching the `DataStoreContext` class,
however the effort to do so would be a non-trivial amount of work
considering this is a temporary feature.

Future work required:
- Mark datastores as tombstone during load when they are tombstoned in
the summary.

Co-authored-by: Navin Agarwal <45832642+agarwal-navin@users.noreply.github.com>
Co-authored-by: Mark Fields <markfields@users.noreply.github.com>