R11s: add throttle cache configs (#12659)
## Description

Currently, throttling logic is set to always use the default values for
the in-memory LRU cache that tracks throttle intervals and status. These
values are too low for production, and we need to be able to configure
them accordingly.

The defaults added to the configs match the current in-code defaults, so
there is no impact to OSS R11s. Also, all configs for throttling remain
optional.